bug 4,6 fix apk compile

+++ app/src/main/java/com/app/onlineshoppersbill/orders/OrderDetailsActivity.java
@@ -164,8 +164,6 @@ public class OrderDetailsActivity extends BaseActivity {
 
                 templatePDF.closeDocument();
                 templatePDF.viewPDF();
-
-
             }
         });
 
@@ -191,7 +189,7 @@ public class OrderDetailsActivity extends BaseActivity {
             qty = orderDetails.get(i).getProductQuantity();
             weight = orderDetails.get(i).getProductWeight();
 
-            costTotal = Integer.parseInt(qty) * Double.parseDouble(price);
+            costTotal = Double.parseDouble(qty) * Double.parseDouble(price);
 
             rows.add(new String[]{name + "\n" + weight + "\n" + "(" + qty + "x" + currency + price + ")", currency + f.format(costTotal)});
